"check_style_score -> 2 of 4 passed. style score: 0/1 "

+++ M3/3-sum/Solution.java
@@ -10,12 +10,12 @@ class Solution {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
 		int size = Integer.parseInt(sc.nextLine());
-		int[] arr = new int[size];
+		long[] arr = new long[size];
 		for (int i = 0; i < size; i++) {
-			arr[i] = Integer.parseInt(sc.nextLine());
+			arr[i] = Long.parseLong(sc.nextLine());
 
 		}
-		int count = 0;
+		long count = 0;
 		for(int i = 0;i<size;i++){
 			for(int j =i+1;j<size;j++){
 				for(int k = j+1;k<size;k++ ){
